+++ derotation/simulate/line_scanning_microscope.py
